+++ src/main/java/com/sprint/discodeit/domain/dto/ProfileImgDto.java
@@ -1,0 +1,4 @@
+package com.sprint.discodeit.domain.dto;
+
+public record ProfileImgDto(String ImgUrl) {
+}
