+++ src/main/java/com/sprint/discodeit/repository/ChannelRepository.java
@@ -1,6 +1,6 @@
 package com.sprint.discodeit.repository;
 
-import com.sprint.discodeit.entity.Channel;
+import com.sprint.discodeit.domain.entity.Channel;
 import com.sprint.discodeit.repository.util.FindRepository;
 import com.sprint.discodeit.repository.util.SaveRepository;
