+++ src/main/java/com/sprint/discodeit/service/ChannelService.java
@@ -2,8 +2,8 @@ package com.sprint.discodeit.service;
 
 
 
-import com.sprint.discodeit.entity.Channel;
-import com.sprint.discodeit.entity.ChannelType;
+import com.sprint.discodeit.domain.entity.Channel;
+import com.sprint.discodeit.domain.entity.ChannelType;
 import java.util.List;
 import java.util.UUID;
